+++ app/components/Views/Snaps/SnapsDev.tsx
@@ -26,7 +26,9 @@ const SnapsDev = () => {
   const navigation = useNavigation();
   const { colors } = useTheme();
 
-  const [snapInput, setSnapInput] = useState<string>('local:http://localhost:3000/snap/');
+  const [snapInput, setSnapInput] = useState<string>(
+    'local:http://localhost:3000/snap/',
+  );
   const snaps = useSelector(
     (state: any) => state.engine.backgroundState.SnapController.snaps,
   );
