+++ app/core/Snaps/location/index.ts
@@ -1,1 +1,2 @@
 export * from './location';
+export * from './fetch';
